challenges/Day 9 Task/ecommerce-agent/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-murf
python-dotenv
